requirements.txt
python-dotenv
openai
pyTelegramBotAPI
flask
